+++ app/src/main/java/edu/uw/tcss450/group4/weatherchatapp/ui/home/RecentChatsFragment.java
@@ -1,4 +1,4 @@
-package edu.uw.tcss450.group4.weatherchatapp;
+package edu.uw.tcss450.group4.weatherchatapp.ui.home;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -12,6 +12,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import edu.uw.tcss450.group4.weatherchatapp.R;
 import edu.uw.tcss450.group4.weatherchatapp.model.PlaceholderContent;
 
 /**
