+++ app/src/main/java/edu/uw/tcss450/group4/weatherchatapp/ui/weather/WeatherViewModel.java
@@ -1,0 +1,4 @@
+package edu.uw.tcss450.group4.weatherchatapp.ui.weather;
+
+public class WeatherViewModel {
+}
