--- app/src/main/java/edu/uw/tcss450/group4/weatherchatapp/weather/WeatherViewModel.java
@@ -1,4 +1,0 @@
-package edu.uw.tcss450.group4.weatherchatapp.weather;
-
-public class WeatherViewModel {
-}
