+++ app/src/main/res/layout/fragment_home.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/weatherFragmentContainer"
-        android:name="edu.uw.tcss450.group4.weatherchatapp.HomeWeatherFragment"
+        android:name="edu.uw.tcss450.group4.weatherchatapp.ui.home.HomeWeatherFragment"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         app:layout_constraintEnd_toEndOf="parent"
@@ -18,7 +18,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/recentChatsFragmentContainer"
-        android:name="edu.uw.tcss450.group4.weatherchatapp.RecentChatsFragment"
+        android:name="edu.uw.tcss450.group4.weatherchatapp.ui.home.RecentChatsFragment"
         android:layout_width="411dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
