+++ app/src/main/res/navigation/main_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="edu.uw.tcss450.group4.weatherchatapp.home.HomeFragment"
+        android:name="edu.uw.tcss450.group4.weatherchatapp.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home"/>
 
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/navigation_weather"
-        android:name="edu.uw.tcss450.group4.weatherchatapp.weather.WeatherFragment"
+        android:name="edu.uw.tcss450.group4.weatherchatapp.ui.weather.WeatherFragment"
         android:label="@string/title_weather"
         tools:layout="@layout/fragment_weather" />
     <fragment
